requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.0
Deprecated==1.2.14
flake8==7.0.0
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.1
google-api-core==2.15.0
google-auth==2.26.1
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
importlib_resources==6.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
nose==1.3.7
ordered-set==4.1.0
packaging==23.2
pinocchio==0.4.3
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
python-dotenv==1.0.0
requests==2.31.0
rich==13.7.1
rsa==4.9
typing_extensions==4.10.0
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
